(+): the PhysicsSystem + WorldPhysics setup is now more flexible and allows easier subclassing.
WorldPhysics is a singleton physics manager which assumed the use of
the Objective-C++ class PhysicsSystem to implement the underlying
fixed-time-step optimised core. PhysicsSystem relied on
CCDraggableSprite as the physics object's Cocos2D counterpart.
Now,  subclassing PhysicsSystem you can redefine the portions relying
on CCDraggableSprite to use a different sprite subclass and enable this
child class in WorldPhysics by subclassing WorldPhysics and its
createWorld* methods only (creating a new children of the PhysicsSystem
subclass and assigning it to WorldPhysics's sharedPhysics's pointer).